requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
bjoern==3.2.2
cffi==1.15.1
click==8.1.3
constantly==15.1.0
cryptography==39.0.2
daphne==4.0.0
Django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
priority==2.0.0
psycogreen==1.0.2
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
python-dotenv==1.0.0
pytz==2022.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
uvicorn==0.21.0
uWSGI==2.0.21
wsproto==1.2.0
zope.interface==5.5.2
